Change DiagramHelper from a model to a class
