+++ src/diagram/index.ts
@@ -1,16 +1,16 @@
 import { Diagram } from "./components/diagram";
 import { EditVariableDialogContent, updateVariable } from "./components/dialog/edit-variable-dialog";
-import { DiagramHelperType } from "./models/diagram-helper";
+import { VariableChip } from "./components/ui/variable-chip";
+import { VariableChipList } from "./components/ui/variable-chip-list";
 import { DQNode } from "./models/dq-node";
 import { DQRoot } from "./models/dq-root";
 import { Variable, VariableType } from "./models/variable";
-import { VariableChip } from "./components/ui/variable-chip";
-import { VariableChipList } from "./components/ui/variable-chip-list";
+import { DiagramHelper } from "./utils/diagram-helper";
 import { useSelectMultipleVariables } from "../hooks/use-select-multiple-variables";
 
 import "./components/diagram.scss";
 
 export {
-  Diagram, DiagramHelperType, EditVariableDialogContent, updateVariable, DQNode, DQRoot, useSelectMultipleVariables,
+  Diagram, DiagramHelper, EditVariableDialogContent, updateVariable, DQNode, DQRoot, useSelectMultipleVariables,
   Variable, VariableChip, VariableChipList, VariableType
 };
